[Fix] 매출 타입 변경
Long -> BigDecimal

Related to: #50

+++ src/main/java/com/midnear/midnearshopping/domain/dto/Statistics/StatisticsDto.java
@@ -14,6 +14,6 @@ import java.util.Date;
 public class StatisticsDto {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date date;
-    private Long paymentAmount;
-    private Long refundAmount;
+    private BigDecimal paymentAmount;
+    private BigDecimal refundAmount;
 }
